about 10% speedup with a manual syscall using inline assembly
The problem is that the calling convention of Linux (SysV) sets all `xmm` registers as clobbered across function calls, and this includes the syscall wrapper of `vmsplice`. By making an inline assembly syscall, I can tell the compiler that the `xmm` register stay fine so that it can preload some more values outside the loop. I hope this still qualifies as "C", as I cannot currently compete with the assembly entries :(